refactor(git-read): use shared search seams
Move bounded search output/process implementations into extensions/shared
so file-search and git-read depend on an owner-neutral module. Add an
import-boundary regression so shared cannot depend on feature extensions.
