+++ extensions/git-read/index.ts
@@ -20,7 +20,7 @@ import type {
 import { Text } from "@earendil-works/pi-tui";
 import { Cause, Effect, Exit } from "effect";
 import { Type } from "typebox";
-import { formatCapturedOutput } from "../file-search/src/output.ts";
+import { formatCapturedOutput } from "../shared/search-output.ts";
 import { sanitizeTerminalText } from "../shared/terminal-text.ts";
 import {
   OPENPI_TOOL_SURFACE,
